Adapt implementation to new RBF structure
